+++ ExtLibs/AltitudeAngelWings/Service/AltitudeAngelService.cs
@@ -313,8 +313,8 @@ namespace AltitudeAngelWings.Service
 
         public PointLatLng newpos(PointLatLng input, double bearing, double distance)
         {
-            const float rad2deg = (float)(180 / Math.PI);
-            const float deg2rad = (float)(1.0 / rad2deg);
+            const double rad2deg = (180 / Math.PI);
+            const double deg2rad = (1.0 / rad2deg);
 
             // '''extrapolate latitude/longitude given a heading and distance 
             //   thanks to http://www.movable-type.co.uk/scripts/latlong.html
