+++ ExtLibs/Controls/HUD.cs
@@ -976,8 +976,8 @@ namespace MissionPlanner.Controls
             }
         }
 
-        const float rad2deg = (float)(180 / Math.PI);
-        const float deg2rad = (float)(1.0 / rad2deg);
+        const double rad2deg = (180 / Math.PI);
+        const double deg2rad = (1.0 / rad2deg);
 
         public void DrawArc(Pen penn, RectangleF rect, float start, float degrees)
         {
