+++ GCSViews/ConfigurationView/ConfigHWUAVCAN.cs
@@ -7,8 +7,8 @@ namespace MissionPlanner.GCSViews.ConfigurationView
 {
     public partial class ConfigHWUAVCAN : UserControl, IActivate
     {
-        private const double rad2deg = (float) (180/Math.PI);
-        private const double deg2rad = (float) (1.0/rad2deg);
+        private const double rad2deg = (180/Math.PI);
+        private const double deg2rad = (1.0/rad2deg);
 
         public ConfigHWUAVCAN()
         {
