+++ Swarm/Formation.cs
@@ -20,8 +20,8 @@ namespace MissionPlanner.Swarm
 
         PointLatLngAlt masterpos = new PointLatLngAlt();
 
-        const double rad2deg = (float)(180 / Math.PI);
-        const double deg2rad = (float)(1.0 / rad2deg);
+        const double rad2deg = (180 / Math.PI);
+        const double deg2rad = (1.0 / rad2deg);
 
         public void setOffsets(MAVState mav, double x, double y, double z)
         {
