+++ Swarm/WaypointLeader/DroneGroup.cs
@@ -17,8 +17,8 @@ namespace MissionPlanner.Swarm.WaypointLeader
     {
         private static readonly ILog log = LogManager.GetLogger(MethodBase.GetCurrentMethod().DeclaringType);
         
-        const double rad2deg = (float)(180 / Math.PI);
-        const double deg2rad = (float)(1.0 / rad2deg);
+        const double rad2deg = (180 / Math.PI);
+        const double deg2rad = (1.0 / rad2deg);
 
         public List<Drone> Drones = new List<Drone>();
         public MAVState airmaster;
