+++ Wizard/6CompassCalib.cs
@@ -138,8 +138,8 @@ namespace MissionPlanner.Wizard
             step++;
         }
 
-        const double rad2deg = (float) (180/Math.PI);
-        const double deg2rad = (float) (1.0/rad2deg);
+        const double rad2deg = (180/Math.PI);
+        const double deg2rad = (1.0/rad2deg);
 
         float calcheading(HIL.Vector3 mag)
         {
